New shared-integration module. - Move SubmissionTestUtil from deposit-integration to shared-integration. - Move/refactor AbstractSubmissionIT from deposit-integration to shared-integration as AbstractSubmissionFixture. - Update the AbstractSubmissionFixture to allow access to the Fedora repository parameters.  Include the base URL of the Fedora repository. - Update existing ITs to compile. - Allow access to the result of the Callable supplied to the Condition on construction.
